facialRecognition-GUI-and-SQL/requirements.txt
mysql-connector-python
pandas
matplotlib
opencv-python
opencv-contrib-python
tk